pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rxo-document-normalizer"
version = "0.1.0"
description = "AI-driven Excel schema standardization platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "azure-functions>=1.20,<2.0",
  "azure-ai-projects>=2.0.0,<3.0",
  "azure-identity>=1.17,<2.0",
  "azure-storage-blob>=12.20,<13.0",
  "openpyxl>=3.1,<4.0",
  "pandas>=2.2,<3.0",
  "pydantic>=2.0,<3.0",
  "streamlit>=1.37,<2.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
